srv/authz/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
structlog==24.4.0
asyncpg==0.29.0
PyJWT==2.9.0
python-dotenv==1.0.1
cryptography==43.0.3





